Fix keymap patches
Quite a few keymap patches incorrectly used 0 as an LED index.

To avoid such mistakes, the existing duplicate entries are now
automatically erased from the keymap while patching.

+++ src/daemon/keymap_patch.h
@@ -1,6 +1,6 @@
 // Keymap patches for specific devices
 typedef struct {
-    int idx; // Index of key to patch in keymap
+    size_t idx; // Index of key to patch in keymap
     const char* name;
     short led;
     short scan;
@@ -11,7 +11,7 @@ typedef struct {
     ushort vendor;
     ushort product;
     keypatch* patch;
-    unsigned patchlen;
+    size_t patchlen;
 } keypatches;
 
 ///
